+++ tests/providers/google/cloud/hooks/test_functions.py
@@ -20,7 +20,7 @@
 import unittest
 
 from airflow import AirflowException
-from airflow.gcp.hooks.functions import CloudFunctionsHook
+from airflow.providers.google.cloud.hooks.functions import CloudFunctionsHook
 from tests.compat import PropertyMock, mock
 from tests.gcp.utils.base_gcp_mock import (
     GCP_PROJECT_ID_HOOK_UNIT_TEST, get_open_mock, mock_base_gcp_hook_default_project_id,
@@ -38,8 +38,8 @@ class TestFunctionHookNoDefaultProjectId(unittest.TestCase):
                         new=mock_base_gcp_hook_no_default_project_id):
             self.gcf_function_hook_no_project_id = CloudFunctionsHook(gcp_conn_id='test', api_version='v1')
 
-    @mock.patch("airflow.gcp.hooks.functions.CloudFunctionsHook._authorize")
-    @mock.patch("airflow.gcp.hooks.functions.build")
+    @mock.patch("airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._authorize")
+    @mock.patch("airflow.providers.google.cloud.hooks.functions.build")
     def test_gcf_client_creation(self, mock_build, mock_authorize):
         result = self.gcf_function_hook_no_project_id.get_conn()
         mock_build.assert_called_once_with(
@@ -53,8 +53,8 @@ class TestFunctionHookNoDefaultProjectId(unittest.TestCase):
         new_callable=PropertyMock,
         return_value=None
     )
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
     def test_create_new_function_missing_project_id(
         self, wait_for_operation_to_complete, get_conn, mock_project_id
     ):
@@ -74,8 +74,8 @@ class TestFunctionHookNoDefaultProjectId(unittest.TestCase):
         self.assertIn("The project id must be passed", str(err))
         wait_for_operation_to_complete.assert_not_called()
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
     def test_create_new_function_overridden_project_id(self, wait_for_operation_to_complete, get_conn):
         create_method = get_conn.return_value.projects.return_value.locations. \
             return_value.functions.return_value.create
@@ -99,7 +99,7 @@ class TestFunctionHookNoDefaultProjectId(unittest.TestCase):
         return_value=None
     )
     @mock.patch('requests.put')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_upload_function_zip_missing_project_id(
         self, get_conn, requests_put, mock_project_id
     ):
@@ -122,7 +122,7 @@ class TestFunctionHookNoDefaultProjectId(unittest.TestCase):
                 self.assertIn("The project id must be passed", str(err))
 
     @mock.patch('requests.put')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_upload_function_zip_overridden_project_id(self, get_conn, requests_put):
         mck, open_module = get_open_mock()
         with mock.patch('{}.open'.format(open_module), mck):
@@ -154,8 +154,8 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
                         new=mock_base_gcp_hook_default_project_id):
             self.gcf_function_hook = CloudFunctionsHook(gcp_conn_id='test', api_version='v1')
 
-    @mock.patch("airflow.gcp.hooks.functions.CloudFunctionsHook._authorize")
-    @mock.patch("airflow.gcp.hooks.functions.build")
+    @mock.patch("airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._authorize")
+    @mock.patch("airflow.providers.google.cloud.hooks.functions.build")
     def test_gcf_client_creation(self, mock_build, mock_authorize):
         result = self.gcf_function_hook.get_conn()
         mock_build.assert_called_once_with(
@@ -169,8 +169,8 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         new_callable=PropertyMock,
         return_value=GCP_PROJECT_ID_HOOK_UNIT_TEST
     )
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
     def test_create_new_function(self, wait_for_operation_to_complete, get_conn, mock_project_id):
         create_method = get_conn.return_value.projects.return_value.locations.\
             return_value.functions.return_value.create
@@ -187,8 +187,8 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         execute_method.assert_called_once_with(num_retries=5)
         wait_for_operation_to_complete.assert_called_once_with(operation_name='operation_id')
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
     def test_create_new_function_override_project_id(self, wait_for_operation_to_complete, get_conn):
         create_method = get_conn.return_value.projects.return_value.locations. \
             return_value.functions.return_value.create
@@ -206,7 +206,7 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         execute_method.assert_called_once_with(num_retries=5)
         wait_for_operation_to_complete.assert_called_once_with(operation_name='operation_id')
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_get_function(self, get_conn):
         get_method = get_conn.return_value.projects.return_value.locations. \
             return_value.functions.return_value.get
@@ -220,8 +220,8 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         get_method.assert_called_once_with(name='function')
         execute_method.assert_called_once_with(num_retries=5)
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
     def test_delete_function(self, wait_for_operation_to_complete, get_conn):
         delete_method = get_conn.return_value.projects.return_value.locations. \
             return_value.functions.return_value.delete
@@ -235,8 +235,8 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         delete_method.assert_called_once_with(name='function')
         execute_method.assert_called_once_with(num_retries=5)
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook._wait_for_operation_to_complete')
     def test_update_function(self, wait_for_operation_to_complete, get_conn):
         patch_method = get_conn.return_value.projects.return_value.locations. \
             return_value.functions.return_value.patch
@@ -263,7 +263,7 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         return_value=GCP_PROJECT_ID_HOOK_UNIT_TEST
     )
     @mock.patch('requests.put')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_upload_function_zip(self, get_conn, requests_put, mock_project_id):
         mck, open_module = get_open_mock()
         with mock.patch('{}.open'.format(open_module), mck):
@@ -288,7 +288,7 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
             )
 
     @mock.patch('requests.put')
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_upload_function_zip_overridden_project_id(self, get_conn, requests_put):
         mck, open_module = get_open_mock()
         with mock.patch('{}.open'.format(open_module), mck):
@@ -313,7 +313,7 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
                 url='http://uploadHere'
             )
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_call_function(self, mock_get_conn):
         payload = {'executionId': 'wh41ppcyoa6l', 'result': 'Hello World!'}
         call = mock_get_conn.return_value.projects.return_value.\
@@ -338,7 +338,7 @@ class TestFunctionHookDefaultProjectId(unittest.TestCase):
         call.assert_called_once_with(body=input_data, name=name)
         self.assertDictEqual(result, payload)
 
-    @mock.patch('airflow.gcp.hooks.functions.CloudFunctionsHook.get_conn')
+    @mock.patch('airflow.providers.google.cloud.hooks.functions.CloudFunctionsHook.get_conn')
     def test_call_function_error(self, mock_get_conn):
         payload = {'error': 'Something very bad'}
         call = mock_get_conn.return_value.projects.return_value. \
