+++ tests/providers/google/cloud/operators/test_functions.py
@@ -24,7 +24,7 @@ from googleapiclient.errors import HttpError
 from parameterized import parameterized
 
 from airflow import AirflowException
-from airflow.gcp.operators.functions import (
+from airflow.providers.google.cloud.operators.functions import (
     FUNCTION_NAME_PATTERN, GcfFunctionDeleteOperator, GcfFunctionDeployOperator, GcfFunctionInvokeOperator,
 )
 from airflow.version import version
@@ -70,7 +70,7 @@ def _prepare_test_bodies():
 
 class TestGcfFunctionDeploy(unittest.TestCase):
     @parameterized.expand(_prepare_test_bodies())
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_body_empty_or_missing_fields(self, body, message, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         with self.assertRaises(AirflowException) as cm:
@@ -84,7 +84,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
         err = cm.exception
         self.assertIn(message, str(err))
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_deploy_execute(self, mock_hook):
         mock_hook.return_value.get_function.side_effect = mock.Mock(
             side_effect=HttpError(resp=MOCK_RESP_404, content=b'not found'))
@@ -111,7 +111,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
             body=expected_body
         )
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_update_function_if_exists(self, mock_hook):
         mock_hook.return_value.get_function.return_value = True
         mock_hook.return_value.update_function.return_value = True
@@ -136,7 +136,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
             expected_body, expected_body.keys())
         mock_hook.return_value.create_new_function.assert_not_called()
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_empty_project_id_is_ok(self, mock_hook):
         mock_hook.return_value.get_function.side_effect = \
             HttpError(resp=MOCK_RESP_404, content=b'not found')
@@ -156,7 +156,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
             location="test_region",
             body=new_body)
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_empty_location(self, mock_hook):
         with self.assertRaises(AirflowException) as cm:
             GcfFunctionDeployOperator(
@@ -168,7 +168,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
         err = cm.exception
         self.assertIn("The required parameter 'location' is missing", str(err))
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_empty_body(self, mock_hook):
         with self.assertRaises(AirflowException) as cm:
             GcfFunctionDeployOperator(
@@ -183,7 +183,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
     @parameterized.expand([
         (runtime,) for runtime in VALID_RUNTIMES
     ])
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_correct_runtime_field(self, runtime, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -207,7 +207,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
             "projects/PRÓJECT/global/networks/netwórk-01"
         ]
     ])
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_valid_network_field(self, network, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -230,7 +230,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
             {"label_324234_a_b_c": 'value-01_93'},
         ]
     ])
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_valid_labels_field(self, labels, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -246,7 +246,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
                                           gcp_conn_id='google_cloud_default')
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_validation_disabled(self, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = {
@@ -265,7 +265,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
                                           gcp_conn_id='google_cloud_default')
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_body_validation_simple(self, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -301,7 +301,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
         ('maxInstances', '-1', "The max instances parameter has to be greater than 0"),
         ('maxInstances', 'ss', "invalid literal for int() with base 10: 'ss'"),
     ])
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_invalid_field_values(self, key, value, message, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -384,7 +384,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
           'repos/b/revisions/c/paths/d'}},
          'test_project_id'),
     ])
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_valid_source_code_union_field(self, source_code, project_id, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         mock_hook.return_value.get_function.side_effect = mock.Mock(
@@ -450,7 +450,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
          "should be of dictionary type")
     ]
     )
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_invalid_trigger_union_field(self, trigger, message, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         body = deepcopy(VALID_BODY)
@@ -486,7 +486,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
                            'service': 'service_name',
                            'failurePolicy': {'retry': {}}}},)
     ])
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_valid_trigger_union_field(self, trigger, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         mock_hook.return_value.get_function.side_effect = mock.Mock(
@@ -515,7 +515,7 @@ class TestGcfFunctionDeploy(unittest.TestCase):
         )
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_extra_parameter(self, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -548,7 +548,7 @@ class TestGcfFunctionDelete(unittest.TestCase):
         'updateTime': '2018-08-23T00:00:00Z',
         'versionId': '1', 'runtime': 'nodejs6'}
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_delete_execute(self, mock_hook):
         mock_hook.return_value.delete_function.return_value = \
             self._DELETE_FUNCTION_EXPECTED
@@ -564,7 +564,7 @@ class TestGcfFunctionDelete(unittest.TestCase):
         )
         self.assertEqual(result['name'], self._FUNCTION_NAME)
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_correct_name(self, mock_hook):
         op = GcfFunctionDeleteOperator(
             name="projects/project_name/locations/project_location/functions"
@@ -575,7 +575,7 @@ class TestGcfFunctionDelete(unittest.TestCase):
         mock_hook.assert_called_once_with(api_version='v1',
                                           gcp_conn_id='google_cloud_default')
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_invalid_name(self, mock_hook):
         with self.assertRaises(AttributeError) as cm:
             op = GcfFunctionDeleteOperator(
@@ -588,7 +588,7 @@ class TestGcfFunctionDelete(unittest.TestCase):
             FUNCTION_NAME_PATTERN))
         mock_hook.assert_not_called()
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_empty_name(self, mock_hook):
         mock_hook.return_value.delete_function.return_value = \
             self._DELETE_FUNCTION_EXPECTED
@@ -601,7 +601,7 @@ class TestGcfFunctionDelete(unittest.TestCase):
         self.assertEqual(str(err), 'Empty parameter: name')
         mock_hook.assert_not_called()
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_gcf_error_silenced_when_function_doesnt_exist(self, mock_hook):
         op = GcfFunctionDeleteOperator(
             name=self._FUNCTION_NAME,
@@ -616,7 +616,7 @@ class TestGcfFunctionDelete(unittest.TestCase):
             'projects/project_name/locations/project_location/functions/function_name'
         )
 
-    @mock.patch('airflow.gcp.operators.functions.CloudFunctionsHook')
+    @mock.patch('airflow.providers.google.cloud.operators.functions.CloudFunctionsHook')
     def test_non_404_gcf_error_bubbled_up(self, mock_hook):
         op = GcfFunctionDeleteOperator(
             name=self._FUNCTION_NAME,
@@ -637,8 +637,8 @@ class TestGcfFunctionDelete(unittest.TestCase):
 
 
 class TestGcfFunctionInvokeOperator(unittest.TestCase):
-    @mock.patch("airflow.gcp.operators.functions.BaseOperator.xcom_push")
-    @mock.patch("airflow.gcp.operators.functions.CloudFunctionsHook")
+    @mock.patch("airflow.providers.google.cloud.operators.functions.BaseOperator.xcom_push")
+    @mock.patch("airflow.providers.google.cloud.operators.functions.CloudFunctionsHook")
     def test_execute(self, mock_gcf_hook, mock_xcom):
         exec_id = 'exec_id'
         mock_gcf_hook.return_value.call_function.return_value = {'executionId': exec_id}
